@@ -4,14 +4,14 @@
 	"license": "Apache",
 	"scripts": {
 		"ng": "ng",
-		"start": "ng serve",
+		"start": "node server.js",
 		"build": "ng build --prod",
 		"gitbuild": "ng build --prod --base-href /dist/",
 		"test": "ng test",
 		"test-ci": "TEST_CI=true ng test",
 		"lint": "ng lint",
 		"e2e": "ng e2e",
-		"heroku-postbuild": "echo Skip build on Heroku"
+		"heroku-postbuild": "ng build --prod"
 	},
 	"private": true,
 	"dependencies": {
@@ -38,12 +38,12 @@
 		"ng2-charts": "1.6.0",
 		"rxjs": "6.3.3",
 		"yarn": "1.12.1",
-		"zone.js": "0.8.26"
+		"zone.js": "0.8.26",
+		"@angular/cli": "^7.2.3",
+		"@angular/compiler-cli": "^7.2.6"
 	},
 	"devDependencies": {
 		"@angular-devkit/build-angular": "^0.12.3",
-		"@angular/cli": "^7.2.3",
-		"@angular/compiler-cli": "^7.2.6",
 		"@angular/language-service": "7.0.2",
 		"@types/jasmine": "2.8.9",
 		"@types/jasminewd2": "2.0.5",
@@ -60,7 +60,7 @@
 		"protractor": "5.4.1",
 		"ts-node": "7.0.1",
 		"tslint": "5.11.0",
-		"typescript": "^3.1.6"
+		"typescript": "^3.2.4"
 	},
 	"heroku-run-build-script": true
 }
