@@ -2,12 +2,12 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
-
 import { AlertModule } from '../shared/modules/alert/alert.module';
 import { HeaderComponent } from './components/header/header.component';
 import { SidebarComponent } from './components/sidebar/sidebar.component';
 import { LayoutRoutingModule } from './layout-routing.module';
 import { LayoutComponent } from './layout.component';
+
 
 @NgModule({
 	imports: [
@@ -17,6 +17,6 @@ import { LayoutComponent } from './layout.component';
 		NgbDropdownModule,
 		AlertModule
 	],
-	declarations: [LayoutComponent, SidebarComponent, HeaderComponent]
+	declarations: [ LayoutComponent, SidebarComponent, HeaderComponent ]
 })
 export class LayoutModule { }
