@@ -1,17 +1,9 @@
 import { User } from './user.model';
 
 export class Income {
-	constructor(
-		public id: number,
+	constructor(public id: number,
 		public user: User,
 		public name: string,
 		public value: number,
-		public date: Date
-	) {
-		this.id = id;
-		this.user = user;
-		this.name = name;
-		this.value = value;
-		this.date = date;
-	}
+		public date: Date) { }
 }
