@@ -1,24 +1,30 @@
 import { routerReducer, RouterReducerState } from '@ngrx/router-store';
 import { ActionReducerMap } from '@ngrx/store';
-
 import { authReducers, AuthState, initialAuthState } from '../auth/store/auth.reducers';
+import { categoryReducers, CategoryState, initialCategoryState } from '../layout/category/store/category.reducers';
+import { expenseReducers, ExpenseState, initialExpenseState } from '../layout/expense/store/expense.reducers';
 import { incomeReducers, IncomeState, initialIncomeState } from '../layout/income/store/income.reducers';
 import { alertReducers, AlertState, initialAlertState } from './alert.reducers';
+
 
 export interface AppState {
 	router?: RouterReducerState;
 	alert: AlertState;
 	auth: AuthState;
 	income: IncomeState;
+	category: CategoryState;
+	expense: ExpenseState;
 }
 
 export const initialAppState: AppState = {
 	alert: initialAlertState,
 	auth: initialAuthState,
-	income: initialIncomeState
+	income: initialIncomeState,
+	category: initialCategoryState,
+	expense: initialExpenseState
 };
 
-export function getInitialState(): AppState {
+export function getInitialState (): AppState {
 	return initialAppState;
 }
 
@@ -26,5 +32,7 @@ export const appReducers: ActionReducerMap<AppState, any> = {
 	router: routerReducer,
 	alert: alertReducers,
 	auth: authReducers,
-	income: incomeReducers
+	income: incomeReducers,
+	category: categoryReducers,
+	expense: expenseReducers
 };
